+++ packages/cli/CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to the Zowe CLI package will be documented in this file.
 
 ## Recent Changes
 
+<<<<<<< HEAD
+- Enhancement: Allow `zowe files view ds ... --range SSS-EEE | SSS,NNN`. [#1539](https://github.com/zowe/zowe-cli/issues/1539)
+=======
 - Enhancement: Added `ZosFilesCreateOptions.alcunit` option to PDS definition. [#1203](https://github.com/zowe/zowe-cli/issues/1203)
+>>>>>>> 27ecd0bf1614165b40182a20322f191568bb6721
 
 ## `7.6.2`